feat: add robots.txt template
J=CR-2868

TEST=manual
