+++ src/templates/robots.tsx
@@ -1,0 +1,28 @@
+import type {
+  TemplateProps,
+  TemplateRenderProps,
+  GetPath,
+  TemplateConfig,
+} from "@yext/pages";
+
+export const config: TemplateConfig = {
+  // The name of the feature. If not set the name of this file will be used (without extension).
+  // Use this when you need to override the feature name.
+  name: "robots",
+};
+
+/**
+ * Defines the path that the generated file will live at for production.
+ *
+ * NOTE: This currently has no impact on the local dev path. Local dev urls currently
+ * take on the form: featureName/entityId
+ */
+export const getPath: GetPath<TemplateProps> = () => {
+  return `robots.txt`;
+};
+
+export function render(data: TemplateRenderProps) {
+  return `User-agent: *
+Sitemap: https://${data.document.siteDomain}/sitemap.xml
+`;
+};
